requirements.txt

Flask
requests
python-dotenv
deep-translator
SpeechRecognition
azure-cognitiveservices-speech
